requirements.txt
eventlet==0.30.2
flask-assets==0.12
flask-migrate==2.1.1
flask-script==2.0.6
flask-socketio==4.2.1
flask-sqlalchemy==2.3.2
flask-wtf
flask>=1.0.2
gunicorn>=19.7.1
markupsafe==2.0.1
psycopg2-binary==2.9.3
pyjwt==2.3.0
python-dotenv==0.12.0
sqlalchemy-utils==0.33.6
sqlalchemy==1.3.24
testing.postgresql==1.3.0
Werkzeug==2.0.3
